LayoutTests/fast/dom/Text/splitText.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test-pre.js"></script>
<script>
description("Tests the splitText API arguments.");

var text = document.createTextNode("abcdefg");

shouldBeEqualToString("text.data", "abcdefg");
shouldBeEqualToString("text.splitText(4).data", "efg");
shouldBeEqualToString("text.data", "abcd");
shouldThrow("text.splitText()", "'TypeError: Not enough arguments'");
shouldThrow("text.splitText(999)", '"IndexSizeError (DOM Exception 1): The index is not in the allowed range."');
shouldThrow("text.splitText(-1)", '"IndexSizeError (DOM Exception 1): The index is not in the allowed range."');
shouldBeEqualToString("text.data", "abcd");
shouldBeEqualToString("text.splitText(-4294967294).data", "cd");
shouldBeEqualToString("text.data", "ab");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
